requirements.txt
pandas
streamlit
streamlit_pandas
os
matplotlib.pyplot
seaborn
plotly.express
plotly.graph_objects
datetime
timedelta
pmdarima 
numpy